requirements/dev.txt
# Base requirements
-r base.txt

# Testing
pytest==7.4.0
pytest-django==4.5.2
pytest-cov==4.1.0
factory-boy==3.3.0
faker==19.6.2

# Linting
flake8==6.1.0
flake8-docstrings==1.7.0
black==23.7.0
isort==5.12.0

# Type checking
mypy==1.5.1
django-stubs==4.2.1

# Pre-commit
pre-commit==3.3.3

# Debugging
django-debug-toolbar==4.2.0
ipython==8.14.0

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==1.3.0 